[Merge] 거래 관련 로직 구현

See merge request s08-webmobile2-sub2/S08P12B111!98

+++ back/src/main/java/com/udteam/miristock/dto/LimitPriceOrderCompareStockDeal.java
@@ -15,14 +15,14 @@ public class LimitPriceOrderCompareStockDeal {
     private String stockCode;
     private String stockName;
     private Integer stockDataDate;
-    private Integer stockDataClosingPrice;
+    private Long stockDataClosingPrice;
     private Long stockDataAmount;
-    private Integer stockDataPriceIncreasement;
+    private Long stockDataPriceIncreasement;
     private Float stockDataFlucauationRate;
     private Integer memberNo;
     private Integer limitPriceOrderNo;
-    private Integer limitPriceOrderPrice;
-    private Integer limitPriceOrderAmount;
+    private Long limitPriceOrderPrice;
+    private Long limitPriceOrderAmount;
     private Deal limitPriceOrderType;
 
     @Builder
